[Silver V] Title: 너의 평점은, Time: 0 ms, Memory: 2088 KB -BaekjoonHub

+++ 백준/Silver/25206. 너의 평점은/너의 평점은.cc
@@ -24,7 +24,7 @@ int main(void) {
 	FAST_IO;
 
 	double total = 0;
-	double total_credit = 0;
+    int total_credit = 0;
 
 	int t = 20;
 	while (t--) {
